requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]>=4.3.3
scikit-learn
dill
pyyaml




##-e .